app/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart>=0.0.9

# RAG stack
langchain==0.3.2
langchain-community==0.3.1
langchain-chroma==0.1.4
chromadb==0.5.17
langchain-text-splitters==0.3.0

# Markdown loading/front-matter
markdown-it-py==3.0.0
python-frontmatter==1.1.0

# File watching for auto reindex
watchdog==5.0.2

# HTTP client
httpx==0.27.2
